Refactor server to allow different handlers (i.e ThreadedServer)
